Work on #65
